+++ gamgee/variant/shared_field_iterator.h
@@ -51,7 +51,7 @@ class SharedFieldIterator : public std::iterator<std::random_access_iterator_tag
    * @param type the encoding of the value 
    * @note this constructor is probably only used by SharedField::begin() and SharedField::end()
    */  
-  explicit SharedFieldIterator(const std::shared_ptr<bcf1_t>& body, uint8_t* data_ptr, uint8_t* end_ptr, const uint8_t bytes_per_value, const utils::VariantFieldType& type) :
+  explicit SharedFieldIterator(const std::shared_ptr<bcf1_t>& body, uint8_t* data_ptr, uint8_t* end_ptr, const uint32_t bytes_per_value, const utils::VariantFieldType& type) :
     m_body {body},
     m_current_data_ptr {data_ptr},
     m_original_data_ptr {data_ptr},
@@ -69,7 +69,7 @@ class SharedFieldIterator : public std::iterator<std::random_access_iterator_tag
    * @param type the encoding of the value 
    * @note this constructor is probably only used by SharedField::begin() and SharedField::end()
    */
-  explicit SharedFieldIterator(const std::shared_ptr<bcf1_t>& body, uint8_t* data_ptr, const uint8_t bytes_per_value, const utils::VariantFieldType& type): SharedFieldIterator(body, data_ptr, nullptr, bytes_per_value, type)
+  explicit SharedFieldIterator(const std::shared_ptr<bcf1_t>& body, uint8_t* data_ptr, const uint32_t bytes_per_value, const utils::VariantFieldType& type): SharedFieldIterator(body, data_ptr, nullptr, bytes_per_value, type)
   {}
    
   SharedFieldIterator(const SharedFieldIterator& other) = default; ///< standard copy constructor creates a new iterator pointing to the same underlying data
@@ -189,7 +189,7 @@ class SharedFieldIterator : public std::iterator<std::random_access_iterator_tag
   const uint8_t* m_current_data_ptr;
   const uint8_t* m_original_data_ptr;
   const uint8_t* m_end_data_ptr;
-  uint8_t m_bytes_per_value;
+  uint32_t m_bytes_per_value;
   utils::VariantFieldType m_type;
   VALUE_TYPE m_current_data_value;
   bool m_is_current_pointee_cached;
